requirement.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
crispy-bootstrap5==0.7
Django==4.1.7
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-filter==22.1
django-js-asset==2.0.0
django-taggit==3.1.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.4.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2023.2
urllib3==1.26.14
